refactor(types): move types shared across app out to shared file
